refactor side-input
